+++ src/one-time-token/controller/exchange.ts
@@ -9,6 +9,7 @@ import * as privilegeService from '../../privilege/service';
 import * as tokenService from '../service';
 import * as oauth2Service from '../../oauth2/service';
 import * as oauth2ClientService from '../../oauth2-client/service';
+import * as userService from '../../user/service';
 
 class OneTimeTokenExchangeController extends Controller {
 
@@ -26,6 +27,16 @@ class OneTimeTokenExchangeController extends Controller {
     }
 
     const user = await tokenService.validateToken(ctx.request.body.token);
+
+    if (!user.active) {
+      if (ctx.request.body.activate) {
+        user.active = true;
+        await userService.save(user);
+      } else {
+        throw new Forbidden('The user associated with the one-time-token has been deactivated. Either activate the user first, or provide the "activate" property in the request if the intent is to activate the user with the one-time-token mechanism');
+      }
+    }
+
     const client = await oauth2ClientService.findByClientId(ctx.request.body.client_id);
     const oauth2Token = await oauth2Service.generateTokenForUser(client, user);